Re-organize files in query_mast module.
