+++ HST/query_mast/__init__.py
@@ -2,50 +2,24 @@
 # query_mast/__init__.py
 ##########################################################################################
 
-import datetime
 import time
-import filecmp
-import fnmatch
 import os
-import re
 
 import pdslogger
-import julian
 
-from astropy.table import Table
-from astropy.table.row import Row
 from astroquery.mast import Observations
 from requests.exceptions import ConnectionError
 
-from product_labels.suffix_info import (ACCEPTED_SUFFIXES,
-                                        ACCEPTED_LETTER_CODES, 
-                                        INSTRUMENT_FROM_LETTER_CODE)
+from .utils import (filter_table, 
+                    is_accepted_instrument_letter_code,
+                    is_accepted_instrument_suffix, 
+                    ymd_tuple_to_mjd)
 
 
 TWD = os.environ["TMP_WORKING_DIR"]
 DEFAULT_DIR = TWD + "/files_from_mast"
 START_DATE = (1900, 1, 1)
 END_DATE = (2025, 1, 1)
-
-def ymd_tuple_to_mjd(ymd):
-    """Return Modified Julian Date.
-    Input:
-        ymd:    a tuple of year, month, and day.
-    """
-    y, m, d = ymd
-    days = julian.day_from_ymd(y, m, d)
-    return julian.mjd_from_day(days)
-
-def filter_table(row_predicate, table):
-    """Return a copy of the filtered table object based on the return of row_predicate.
-    Input:
-        row_predicate:    a function with the condition used to filter the table.
-        table:            target table to be filtered.
-    """
-    to_delete = [n for (n, row) in enumerate(table) if not row_predicate(row)]
-    copy = table.copy()
-    copy.remove_rows(to_delete)
-    return copy
 
 def get_products_from_mast(proposal_id,
                            start_date=START_DATE,
@@ -88,7 +62,6 @@ def get_products_from_mast(proposal_id,
     # Download all accepted files
     download_files(filtered_products, proposal_id, logger, dir)
 
-
 def query_mast_slice(proposal_id,
                      start_date=START_DATE,
                      end_date=END_DATE,
@@ -126,6 +99,17 @@ def query_mast_slice(proposal_id,
     logger.exception(RuntimeError)
     raise RuntimeError("Query mast timed out. Number of retries: " + max_retries)
 
+def get_filtered_products(table):
+    """Return product rows of an observation table with accepted instrument letter code
+    and suffxes.
+    Input:
+        table:  an observation table from mast query.
+    """
+    result = Observations.get_product_list(table)
+    result = filter_table(is_accepted_instrument_letter_code, result)
+    result = filter_table(is_accepted_instrument_suffix, result)
+    return result
+
 def download_files(table, proposal_id, logger=None, dir=DEFAULT_DIR):
     """Download files from mast for a given product table and proposal id
     Input:
@@ -144,50 +128,3 @@ def download_files(table, proposal_id, logger=None, dir=DEFAULT_DIR):
     if len(table) > 0:
         logger.info("Download files to " + working_dir)
         Observations.download_products(table, download_dir=working_dir)
-
-def get_filtered_products(table):
-    """Return product rows of an observation table with accepted instrument letter code
-    and suffxes.
-    Input:
-        table:  an observation table from mast query.
-    """
-    result = Observations.get_product_list(table)
-    result = filter_table(is_accepted_instrument_letter_code, result)
-    result = filter_table(is_accepted_instrument_suffix, result)
-    # print(ACCEPTED_SUFFIXES)
-    return result
-
-
-def is_accepted_instrument_letter_code(row):
-    """Check if a product row has accepted letter code in the first letter of the 
-    product filename.
-    Input:
-        row:    an observation table row.
-    """
-    return row["obs_id"][0].lower() in ACCEPTED_LETTER_CODES
-
-def is_accepted_instrument_suffix(row):
-    """Check if a product row has accepted suffex in the productSubGroupDescription field
-    of the table.
-    Input:
-        row:    an observation table row.
-    """
-    suffix = get_suffix(row)
-    instrument_id = get_instrument_id(row)
-    # For files like n4wl03fxq_raw.jpg with "--" will raise an error
-    # return is_accepted(suffix, instrument_id) 
-    return suffix in ACCEPTED_SUFFIXES[instrument_id]
-
-def get_instrument_id(row):
-    """Return the instrument id for a given product row.
-    Input:
-        row:    an observation table row.
-    """
-    return INSTRUMENT_FROM_LETTER_CODE[row["obs_id"][0].lower()]
-
-def get_suffix(row):
-    """Return the product file suffix for a given product row.
-    Input:
-        row:    an observation table row.
-    """
-    return str(row["productSubGroupDescription"]).lower()
